Move property tests to another module
