+++ pallets/gas/src/lib.rs
@@ -31,6 +31,8 @@ mod mock;
 
 #[cfg(test)]
 mod tests;
+#[cfg(test)]
+mod property_tests;
 
 #[derive(Clone, Decode, Debug, Encode, MaxEncodedLen, TypeInfo)]
 pub enum ValueType {
